site/requirements.txt
asgiref==3.2.7
Django==3.0.6
gunicorn==20.0.4
humanhash3==0.0.6
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
